fix: ensure `preferred_dtype` is correct for `Count`

+++ src/awkward/_connect/jax/reducers.py
@@ -38,7 +38,7 @@ class ArgMax(Reducer):
 
 class Count(Reducer):
     name = "count"
-    preferred_dtype = np.float64
+    preferred_dtype = np.int64
 
     @classmethod
     def return_dtype(cls, given_dtype):
@@ -53,7 +53,7 @@ class Count(Reducer):
 
 class CountNonzero(Reducer):
     name = "count_nonzero"
-    preferred_dtype = np.float64
+    preferred_dtype = np.int64
 
     @classmethod
     def return_dtype(cls, given_dtype):
